// clients/ts/src/generated/types/paymentStatus.ts
/**
 * This code was AUTOGENERATED using the Codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun Codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import {
  combineCodec,
  getEnumDecoder,
  getEnumEncoder,
  type FixedSizeCodec,
  type FixedSizeDecoder,
  type FixedSizeEncoder,
} from "@solana/kit";

export enum PaymentStatus {
  Created,
  Acknowledged,
  Settled,
  Cancelled,
  Expired,
}

export type PaymentStatusArgs = PaymentStatus;

export function getPaymentStatusEncoder(): FixedSizeEncoder<PaymentStatusArgs> {
  return getEnumEncoder(PaymentStatus);
}

export function getPaymentStatusDecoder(): FixedSizeDecoder<PaymentStatus> {
  return getEnumDecoder(PaymentStatus);
}

export function getPaymentStatusCodec(): FixedSizeCodec<
  PaymentStatusArgs,
  PaymentStatus
> {
  return combineCodec(getPaymentStatusEncoder(), getPaymentStatusDecoder());
}